updated the action structs with the correct schemas that the game server needs and also made a client package struct where all the client data will be sent

+++ shared/objects.go
@@ -52,7 +52,7 @@ func (BombData) isObjectData() {}
 type Object struct {
 	Type       ObjectType
 	Data       *any
-	Id         uint64
+	Id         uint
 	Pos        Position
 	Hp         int32
 	ObjectData ObjectData
